reorganize vector interaction code
